requirements.txt
# ── STT ──────────────────────────────────────────────
faster-whisper==1.0.3
sounddevice==0.4.6
numpy==1.26.4

# ── Agent (LangGraph + LangChain) ────────────────────
langchain==0.2.16
langgraph==0.2.28
langchain-openai==0.1.23
langchain-groq==0.1.9
groq==0.9.0

# ── Vision & Camera ───────────────────────────────────
opencv-python==4.10.0.84
Pillow==10.4.0

# ── TTS ───────────────────────────────────────────────
gTTS==2.5.3
playsound==1.2.2
elevenlabs==1.9.0

# ── Utilities ─────────────────────────────────────────
python-dotenv==1.0.1
loguru==0.7.2
